+++ ktlint/src/main/kotlin/com/pinterest/ktlint/internal/FileUtils.kt
@@ -1,37 +1,116 @@
 package com.pinterest.ktlint.internal
 
-import com.github.shyiko.klob.Glob
 import com.pinterest.ktlint.core.KtLint
 import com.pinterest.ktlint.core.LintError
 import com.pinterest.ktlint.core.RuleSet
 import java.io.File
+import java.nio.file.FileSystem
+import java.nio.file.FileVisitResult
+import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
+import java.nio.file.SimpleFileVisitor
+import java.nio.file.attribute.BasicFileAttributes
 import kotlin.system.exitProcess
 
 internal val workDir: String = File(".").canonicalPath
+private val tildeRegex = Regex("^(!)?~")
 
-internal fun List<String>.fileSequence(): Sequence<File> {
-    val kotlinFiles: Sequence<Path> = if (isEmpty()) {
-        Glob.from("**/*.kt", "**/*.kts")
-            .iterate(
-                Paths.get(workDir),
-                Glob.IterationOption.SKIP_HIDDEN
-            )
-            .asSequence()
+internal fun FileSystem.fileSequence(
+    globs: List<String>,
+    rootDir: Path = Paths.get(".").toAbsolutePath().normalize()
+): Sequence<Path> {
+    checkGlobsContainAbsolutePath(globs)
+
+    val pathMatchers = if (globs.isEmpty()) {
+        setOf(
+            getPathMatcher("glob:**$globSeparator*.kt"),
+            getPathMatcher("glob:**$globSeparator*.kts")
+        )
     } else {
-        // Converting List<String> to Array<String> and passing it to Glob.from(patterns) skips some files
-        // See https://github.com/pinterest/ktlint/issues/942
-        map {
-            Glob.from(expandTilde(it))
-                .iterate(Paths.get(workDir))
-        }
-            .asSequence()
-            .flatMap { it.asSequence() }
+        globs
+            .filterNot { it.startsWith("!") }
+            .map {
+                getPathMatcher(it.toGlob(rootDir))
+            }
     }
 
-    return kotlinFiles
-        .map(Path::toFile)
+    val negatedPathMatchers = if (globs.isEmpty()) {
+        emptySet()
+    } else {
+        globs
+            .filter { it.startsWith("!") }
+            .map { getPathMatcher(it.removePrefix("!").toGlob(rootDir)) }
+    }
+
+    val result = mutableListOf<Path>()
+    Files.walkFileTree(
+        rootDir,
+        object : SimpleFileVisitor<Path>() {
+            override fun visitFile(
+                filePath: Path,
+                fileAttrs: BasicFileAttributes
+            ): FileVisitResult {
+                if (negatedPathMatchers.none { it.matches(filePath) } &&
+                    pathMatchers.any { it.matches(filePath) }
+                ) {
+                    result.add(filePath)
+                }
+                return FileVisitResult.CONTINUE
+            }
+
+            override fun preVisitDirectory(
+                dirPath: Path,
+                dirAttr: BasicFileAttributes
+            ): FileVisitResult {
+                return if (Files.isHidden(dirPath)) {
+                    FileVisitResult.SKIP_SUBTREE
+                } else {
+                    FileVisitResult.CONTINUE
+                }
+            }
+        }
+    )
+
+    return result.asSequence()
+}
+
+private fun FileSystem.checkGlobsContainAbsolutePath(globs: List<String>) {
+    val rootDirs = rootDirectories.map { it.toString() }
+    globs
+        .map { it.removePrefix("!") }
+        .filter { glob ->
+            rootDirs.any { glob.startsWith(it) }
+        }
+        .run {
+            if (isNotEmpty()) {
+                throw IllegalArgumentException(
+                    "KtLint does not support absolute path in globs:\n${joinToString(separator = "\n")}"
+                )
+            }
+        }
+}
+
+private fun String.toGlob(
+    rootDir: Path
+): String {
+    val expandedPath = expandTilde(this)
+    val rootDirPath = rootDir
+        .toAbsolutePath()
+        .toString()
+        .run {
+            val normalizedPath = if (!endsWith(File.separator)) "$this${File.separator}" else this
+            normalizedPath.replace(File.separator, globSeparator)
+        }
+    return "glob:$rootDirPath$expandedPath"
+}
+
+internal val globSeparator: String get() {
+    val os = System.getProperty("os.name")
+    return when {
+        os.startsWith("windows", ignoreCase = true) -> "\\\\"
+        else -> "/"
+    }
 }
 
 /**
@@ -50,7 +129,7 @@ internal fun JarFiles.toFilesURIList() = map {
 
 // a complete solution would be to implement https://www.gnu.org/software/bash/manual/html_node/Tilde-Expansion.html
 // this implementation takes care only of the most commonly used case (~/)
-private fun expandTilde(path: String): String = path.replaceFirst(Regex("^~"), System.getProperty("user.home"))
+private fun expandTilde(path: String): String = path.replaceFirst(tildeRegex, System.getProperty("user.home"))
 
 internal fun File.location(
     relative: Boolean